internal/util: Rename from `internal/misc`.
