+++ internal/util/util.go
@@ -1,4 +1,4 @@
-package misc
+package util
 
 import (
 	"io"
